seventh 1

+++ 0x15-file_io/1-create_file.c
@@ -14,7 +14,7 @@
 
 int create_file(const char *filename, char *text_content)
 {
-	int descriptor;
+	ssize_t descriptor;
 	ssize_t num_of_bytes = 0;
 	int length = strlen(text_content);
